+++ crates/ruma-events/src/call.rs
@@ -6,12 +6,8 @@ pub mod answer;
 pub mod candidates;
 pub mod hangup;
 pub mod invite;
-#[cfg(feature = "unstable-msc3401")]
-pub mod matrix_rtc;
-
 pub mod negotiate;
-#[cfg(feature = "unstable-msc4075")]
-pub mod notify;
+
 pub mod reject;
 #[cfg(feature = "unstable-msc3291")]
 pub mod sdp_stream_metadata_changed;
